Add api-client tsconfig + implement comments
