+++ tsconfig.json
@@ -20,6 +20,7 @@
       "@screens/*": ["./screens/*"],
       "@screens": ["./screens"],
       "@components/*": ["./components/*"],
+      "@api-client/*": ["./api-client/*"],
       "@api-client": ["./api-client"],
       "@stacks/*": ["./stacks/*"],
       "@hooks/*": ["./hooks/*"],
